Add GetProperty C function for KOD, nomoveon.kod changes.
Added a new C function for KOD to call, named GetProperty. As the name
suggests, its function is to return the contents of a property given an
object and a property string.

Form is GetProperty(object,"property"). This is much faster to run than
Send(object,@GetProperty).

For testing and examples, I've changed some of the Send calls for
properties in nomoveon.kod to GetProperty calls. On my computer I time
SqFineDistToVision at 1290ms for 400k calls with Send, and 933ms for
400k calls with GetProperty. IsInSameRoom took 2128ms for 400k calls
with Send, 1103ms for 400k calls with GetProperty (and 533ms with
Send(self,@GetOwner) changed to poOwner).