Force Spike
Force Spike allows a mage to expend a huge quantity of mana to power
through a fizzle. Amount expended starts at triple the cost of the
spell, then relies on spellpower, skill in the skill and spell, and
reaches a minimum at 1x the cost of the spell (paid a second time).