Add C_AppendString C function for adding to TAG_STRING.
This C function takes a string of type TAG_STRING or $ for the first
parameter, and any number of following parameters and appends them in
order to the string. If passed $, creates a new string. Returns the int
value of the old or created string.

Can accept data of type TAG_INT, TAG_STRING, TAG_DEBUGSTR, TAG_MESSAGE,
TAG_TEMP_STRING, TAG_RESOURCE and notably, TAG_LIST. If passed a list,
will add each element of the list in order (if valid data) to the list.
This means a list containing objects (which we don't want) and strings
(which we do) would have just the strings added.